fix(events): globally-unique u128 event ids (P2.1)
EventMetadata::event_id was a process-local AtomicU64: two instances of
the same app generated colliding ids, so one poller could silently drop
a peer's message from the emit_and_wait dedup set. The id is now a u128
composed of a per-process random identity (high 64 bits, drawn from the
OS CSPRNG via the existing uuid workspace dep) and the per-process
counter (low 64 bits). Codec parses u128; dedup set keys on u128.
Breaking change to EventMetadata (pre-production).

+++ r2e-events/src/backend/metadata_codec.rs
@@ -46,7 +46,7 @@ pub fn decode_metadata(
         let v = value.as_ref();
         match k {
             HEADER_EVENT_ID => {
-                if let Ok(id) = v.parse::<u64>() {
+                if let Ok(id) = v.parse::<u128>() {
                     metadata.event_id = id;
                 }
             }
